messagepack: perform full reset
